review: split the credit-cap discovery and clamp into credit-cap.ts
router.ts crossed the 1000-line lint budget with the clamp addition; the
cap discovery (resolveCreditCap), the clamp (clampBudgetToCreditCap), and
the gh-aw default constant move to their own module by concern. The router
imports both and stays the single CLI entry point; tests for the moved
functions import from the new module. Prettier fix in router.test.ts rides
along.
